Add new account

+++ app/src/components/Logout/functions/logoutUser.js
@@ -1,4 +1,10 @@
 export const logoutUser = async (token) => {
+  if (!token) {
+    return Promise.reject({
+      errorCode: "invalid-params",
+      errorMessage: "Missing token",
+    });
+  }
   const endpoint = process.env.REACT_APP_API_ENDPOINT;
   return fetch(endpoint + "sessions", {
     method: "DELETE",
